chore: feature connected to data base. Added mapper to avoid changes in types to spread
