+++ app/__tests__/compare.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react';
-import Compare from '../compare/page';
+import ComparePage from '../compare/page';
 
 describe('Compare Page', () => {
   it('renders hello', () => {
-    render(<Compare />);
+    render(<ComparePage />);
     const compareText = screen.getAllByText(/SELECT A CHARACTER/i);
     expect(compareText.length).toBe(2);
   });
